memory: mtk: Fix incompatible pointer to integer conversion
error: incompatible pointer to integer conversion assigning to 'unsigned long' from 'void *' [-Wint-conversion]
        mpucb->owner = __builtin_return_address(0);
                     ^ ~~~~~~~~~~~~~~~~~~~~~~~~~~~

+++ drivers/memory/mediatek/emimpu.c
@@ -953,7 +953,7 @@ int mtk_emimpu_register_callback(
 	if (!mpucb)
 		return -ENOMEM;
 
-	mpucb->owner = __builtin_return_address(0);
+	mpucb->owner = (unsigned long)__builtin_return_address(0);
 	mpucb->debug_dump = debug_dump;
 	mpucb->handled = false;
